// src/api/generated/model/groupAddressDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Chargee Ampere
 * API v2
 * OpenAPI spec version: 2.0.0
 */
import type { GroupAddressDtoSparky } from './groupAddressDtoSparky';
import type { GroupAddressDtoFlint } from './groupAddressDtoFlint';

export interface GroupAddressDto {
  /** The UUID of the address */
  uuid: string;
  /**
   * The sparky associated with the address
   * @nullable
   */
  sparky: GroupAddressDtoSparky;
  /**
   * The flint associated with the address
   * @nullable
   */
  flint: GroupAddressDtoFlint;
  /** When the device was added to the group */
  createdAt: string;
  /**
   * When the group-to-device relation was last updated
   * @nullable
   */
  updatedAt: string | null;
}