// src/api/generated/model/smartMeterDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Chargee Ampere
 * API v2
 * OpenAPI spec version: 2.0.0
 */

export interface SmartMeterDto {
  /** The uuid of the smart meter */
  identifier: string;
  /** The uuid of the address this smart meter belongs to */
  addressUuid: string;
  /** The type of the smart meter */
  smartMeterType: string;
  /** The meter number of the smart meter */
  meterNumber: string;
  /** The gas meter number of the smart meter */
  gasMeterNumber: string;
  /**
   * The electricity EAN of the smart meter
   * @nullable
   */
  eanElectricity: string | null;
  /**
   * The gas EAN of the smart meter
   * @nullable
   */
  eanGas: string | null;
  /** Created at timestamp */
  createdAt: string;
  /**
   * Updated at timestamp
   * @nullable
   */
  updatedAt: string | null;
}